divide project on modules
