+++ js/random-number.js
@@ -1,0 +1,6 @@
+const getRandomNumber = function (min, max, decimalPoints) {
+  const findRandom = Math.random() * (max - min) + min;
+  return +findRandom.toFixed(decimalPoints);
+};
+
+export { getRandomNumber };
